update animation for knight
The test is in BattleTestAcitivity. I added a new class, PlayerImage in
package com.example.longdungeon.layout, and changed it in battle test
layout. The commit is messy because i tested many functions to make it
work, and you can ignore it. Just look on the class PlayerImage,
BattleTestActivity layout, and LightAttack from ListView.